restructure smatrix
