+++ src/smatrix/blaslapack/eig.rs
@@ -7,10 +7,9 @@
 //--------------------------------------------------------------------------------------------------
 
 //{{{ crate imports
-use super::SMatrix;
-use crate::blaslapack::common::AsI32;
-use crate::blaslapack::geev::eig_raw;
-use crate::blaslapack::geev::{self, Geev};
+use crate::smatrix::SMatrix;
+use crate::blaslapack::AsI32;
+use crate::blaslapack::{eig_raw, EigRawError, Geev};
 use crate::common::{Complex, Field, One, Zero};
 //}}}
 //{{{ dep imports
@@ -25,7 +24,7 @@ pub enum Error
 {
     /// Wraps a LAPACK `geev` error from the eigenvalue computation routine.
     #[error("Error in eig(), exited with error:\n{0}")]
-    GeevError(#[from] geev::Error),
+    GeevError(#[from] EigRawError),
 }
 //}}}
