+++ src/smatrix/elementwise/addop.rs
@@ -1,26 +1,29 @@
-//! Element-wise division operators for [`SMatrix`]: matrix / scalar and matrix / matrix.
+//! Addition operators for [`SMatrix`]: matrix + matrix and matrix + scalar.
 //!
-//! Implements the [`Div`] trait for element-wise (Hadamard) division of [`SMatrix<T, N, M>`]
-//! operands and for matrix–scalar scaling. Like `mulop`, this performs element-wise division
-//! rather than any matrix-theoretic inverse. Results are lazy `BinopExpr` values; shape
-//! compatibility is enforced at compile time through the const-generic dimension parameters.
+//! Implements the standard [`Add`] trait for all combinations of owned and borrowed
+//! [`SMatrix<T, N, M>`] operands, as well as mixed matrix–scalar and scalar–matrix overloads.
+//! Matrix–matrix addition returns a lazy `BinopExpr` that defers allocation until materialised;
+//! scalar addition is applied element-wise. Const-generic dimension parameters ensure that only
+//! shape-compatible matrices can be added at compile time.
 //--------------------------------------------------------------------------------------------------
 
 //{{{ crate imports
-use super::SMatrix;
+use crate::smatrix::SMatrix;
 use crate::apply_for_all_types;
 #[cfg(feature = "enable_checks")]
 use crate::common::Shape;
 use crate::common::{Field, IndexValue, LazyExpr};
-use crate::expression::binary_expr::{BinopExpr, DivOp};
+use crate::expression::binary_expr::{AddOp, BinopExpr};
 //}}}
 //{{{ std imports
-use std::ops::{Div, DivAssign};
+use std::ops::{Add, AddAssign};
+//}}}
+//{{{ dep imports
 //}}}
 //--------------------------------------------------------------------------------------------------
 //{{{ collection: eagerly evaluated expressions
-//{{{ impl: Div<T> for SMatrix
-impl<T, const N: usize, const M: usize> Div<T> for SMatrix<T, N, M>
+//{{{ impl: Add<T> for SMatrix
+impl<T, const N: usize, const M: usize> Add<T> for SMatrix<T, N, M>
 where
     [(); N * M]:,
     T: Field + Copy,
@@ -28,19 +31,19 @@ where
     type Output = SMatrix<T, N, M>;
 
     #[inline]
-    fn div(
+    fn add(
         self,
         rhs: T,
     ) -> Self::Output
     {
         let mut out = self;
-        out.iter_mut().for_each(|x| *x /= rhs);
+        out.iter_mut().for_each(|x| *x += rhs);
         out
     }
 }
 //}}}
-//{{{ impl: Div<SMatrix> for SMatrix
-impl<T, const N: usize, const M: usize> Div for SMatrix<T, N, M>
+//{{{ impl: Add<SMatrix> for SMatrix
+impl<T, const N: usize, const M: usize> Add for SMatrix<T, N, M>
 where
     [(); N * M]:,
     T: Field + Copy + IndexValue<usize, Output = T>,
@@ -48,7 +51,7 @@ where
     type Output = SMatrix<T, N, M>;
 
     #[inline]
-    fn div(
+    fn add(
         self,
         rhs: SMatrix<T, N, M>,
     ) -> Self::Output
@@ -64,62 +67,62 @@ where
         out.iter_mut()
             .zip(rhs.iter())
             .for_each(|(out_elem, rhs_elem)| {
-                *out_elem /= *rhs_elem;
+                *out_elem += *rhs_elem;
             });
 
         out
     }
 }
 //}}}
-//{{{ impl Div<SMatrix<T, N, M>> for T
-macro_rules! impl_smatrix_div_owned {
+//{{{ impl Add<SMatrix<T, N, M>> for T
+macro_rules! impl_smatrix_add_owned {
     ($type: ty) => {
         #[doc(hidden)]
-        impl<const N: usize, const M: usize> Div<SMatrix<$type, N, M>> for $type
+        impl<const N: usize, const M: usize> Add<SMatrix<$type, N, M>> for $type
         where
             [(); N * M]:,
         {
             type Output = SMatrix<$type, N, M>;
 
             #[inline]
-            fn div(
+            fn add(
                 self,
                 rhs: SMatrix<$type, N, M>,
             ) -> Self::Output
             {
                 let mut out = rhs;
-                out.iter_mut().for_each(|x| *x = self / *x);
+                out.iter_mut().for_each(|x| *x += self);
                 out
             }
         }
     };
 }
-apply_for_all_types!(impl_smatrix_div_owned);
+apply_for_all_types!(impl_smatrix_add_owned);
 //}}}
-//{{{ impl DivAssign<T> for SMatrix
-impl<T, const N: usize, const M: usize> DivAssign<T> for SMatrix<T, N, M>
+//{{{ impl AddAssign<T> for SMatrix
+impl<T, const N: usize, const M: usize> AddAssign<T> for SMatrix<T, N, M>
 where
     [(); N * M]:,
     T: Field + Copy,
 {
     #[inline]
-    fn div_assign(
+    fn add_assign(
         &mut self,
         rhs: T,
     )
     {
-        self.iter_mut().for_each(|x| *x /= rhs);
+        self.iter_mut().for_each(|x| *x += rhs);
     }
 }
 //}}}
-//{{{ impl: DivAssign<SMatrix> for SMatrix
-impl<T, const N: usize, const M: usize> DivAssign for SMatrix<T, N, M>
+//{{{ impl: AddAssign<SMatrix> for SMatrix
+impl<T, const N: usize, const M: usize> AddAssign for SMatrix<T, N, M>
 where
     [(); N * M]:,
     T: Field + Copy,
 {
     #[inline]
-    fn div_assign(
+    fn add_assign(
         &mut self,
         rhs: SMatrix<T, N, M>,
     )
@@ -134,31 +137,31 @@ where
         self.iter_mut()
             .zip(rhs.iter())
             .for_each(|(out_elem, rhs_elem)| {
-                *out_elem /= *rhs_elem;
+                *out_elem += *rhs_elem;
             });
     }
 }
 //}}}
 //}}}
-//{{{ collection: DivOp for SMatrix
-//{{{ impl: Div<T> for SMatrix
-macro_rules! impl_smatrix_div_scalar_rhs {
+//{{{ collection: AddOp for SMatrix
+//{{{ impl: Add<T> for SMatrix
+macro_rules! impl_smatrix_add_scalar_rhs {
     ($type:ty) => {
         #[doc(hidden)]
-        impl<'a, const N: usize, const M: usize> Div<$type> for &'a SMatrix<$type, N, M>
+        impl<'a, const N: usize, const M: usize> Add<$type> for &'a SMatrix<$type, N, M>
         where
             [(); N * M]:,
         {
-            type Output = BinopExpr<&'a SMatrix<$type, N, M>, $type, $type, DivOp>;
+            type Output = BinopExpr<&'a SMatrix<$type, N, M>, $type, $type, AddOp>;
 
-            fn div(
+            #[inline]
+            fn add(
                 self,
                 rhs: $type,
             ) -> Self::Output
             {
                 let nr = self.nrows;
                 let nc = self.ncols;
-
                 BinopExpr {
                     a: self,
                     b: rhs,
@@ -170,45 +173,45 @@ macro_rules! impl_smatrix_div_scalar_rhs {
         }
 
         #[doc(hidden)]
-        impl<'a, const N: usize, const M: usize> Div<$type> for &'a mut SMatrix<$type, N, M>
+        impl<'a, const N: usize, const M: usize> Add<$type> for &'a mut SMatrix<$type, N, M>
         where
             [(); N * M]:,
         {
-            type Output = BinopExpr<&'a SMatrix<$type, N, M>, $type, $type, DivOp>;
+            type Output = BinopExpr<&'a SMatrix<$type, N, M>, $type, $type, AddOp>;
 
             #[inline]
-            fn div(
+            fn add(
                 self,
                 rhs: $type,
             ) -> Self::Output
             {
-                (&*self).div(rhs)
+                (&*self).add(rhs)
             }
         }
     };
 }
 
-apply_for_all_types!(impl_smatrix_div_scalar_rhs);
+apply_for_all_types!(impl_smatrix_add_scalar_rhs);
 
 //}}}
-//{{{ impl: Div<Smatrix> for $type
-macro_rules! impl_smatrix_div {
+//{{{ impl: Add<Smatrix> for $type
+macro_rules! impl_smatrix_add {
     ($type:ty) => {
         #[doc(hidden)]
-        impl<'a, const N: usize, const M: usize> Div<&'a SMatrix<$type, N, M>> for $type
+        impl<'a, const N: usize, const M: usize> Add<&'a SMatrix<$type, N, M>> for $type
         where
             [(); N * M]:,
         {
-            type Output = BinopExpr<$type, &'a SMatrix<$type, N, M>, $type, DivOp>;
+            type Output = BinopExpr<$type, &'a SMatrix<$type, N, M>, $type, AddOp>;
 
-            fn div(
+            #[inline]
+            fn add(
                 self,
                 rhs: &'a SMatrix<$type, N, M>,
             ) -> Self::Output
             {
                 let nr = rhs.nrows;
                 let nc = rhs.ncols;
-
                 BinopExpr {
                     a: self,
                     b: rhs,
@@ -221,44 +224,44 @@ macro_rules! impl_smatrix_div {
     };
 }
 
-apply_for_all_types!(impl_smatrix_div);
+apply_for_all_types!(impl_smatrix_add);
 //}}}
-//{{{ impl: Div<&mut Smatrix> for $type
-macro_rules! impl_smatrix_div_mut {
+//{{{ impl: Add<&mut Smatrix> for $type
+macro_rules! impl_smatrix_add_mut {
     ($type:ty) => {
         #[doc(hidden)]
-        impl<'a, const N: usize, const M: usize> Div<&'a mut SMatrix<$type, N, M>> for $type
+        impl<'a, const N: usize, const M: usize> Add<&'a mut SMatrix<$type, N, M>> for $type
         where
             [(); N * M]:,
         {
-            type Output = BinopExpr<$type, &'a SMatrix<$type, N, M>, $type, DivOp>;
+            type Output = BinopExpr<$type, &'a SMatrix<$type, N, M>, $type, AddOp>;
 
             #[inline]
-            fn div(
+            fn add(
                 self,
                 rhs: &'a mut SMatrix<$type, N, M>,
             ) -> Self::Output
             {
-                self.div(&*rhs)
+                self.add(&*rhs)
             }
         }
     };
 }
 
-apply_for_all_types!(impl_smatrix_div_mut);
+apply_for_all_types!(impl_smatrix_add_mut);
 
 //}}}
-//{{{ impl: Div<Rhs> for &'a SMatrix
-impl<'a, T, Rhs, const N: usize, const M: usize> Div<Rhs> for &'a SMatrix<T, N, M>
+//{{{ impl: Add<Rhs> for &'a SMatrix
+impl<'a, T, Rhs, const N: usize, const M: usize> Add<Rhs> for &'a SMatrix<T, N, M>
 where
     [(); N * M]:,
     T: Field + Copy,
     Rhs: LazyExpr<ScalarType = T> + IndexValue<usize, Output = T>,
 {
-    type Output = BinopExpr<&'a SMatrix<T, N, M>, Rhs, T, DivOp>;
+    type Output = BinopExpr<&'a SMatrix<T, N, M>, Rhs, T, AddOp>;
 
     #[inline]
-    fn div(
+    fn add(
         self,
         rhs: Rhs,
     ) -> Self::Output
@@ -282,22 +285,22 @@ where
 }
 
 //}}}
-//{{{ impl: Div<Rhs> for &'a mut SMatrix
-impl<'a, T, Rhs, const N: usize, const M: usize> Div<Rhs> for &'a mut SMatrix<T, N, M>
+//{{{ impl: Add<Rhs> for &'a mut SMatrix
+impl<'a, T, Rhs, const N: usize, const M: usize> Add<Rhs> for &'a mut SMatrix<T, N, M>
 where
     [(); N * M]:,
     T: Field + Copy,
     Rhs: LazyExpr<ScalarType = T> + IndexValue<usize, Output = T>,
 {
-    type Output = BinopExpr<&'a SMatrix<T, N, M>, Rhs, T, DivOp>;
+    type Output = BinopExpr<&'a SMatrix<T, N, M>, Rhs, T, AddOp>;
 
     #[inline]
-    fn div(
+    fn add(
         self,
         rhs: Rhs,
     ) -> Self::Output
     {
-        (&*self).div(rhs)
+        (&*self).add(rhs)
     }
 }
